feat: add support for gnome-terminal and konsole and some refactors
- In order to support gnome-terminal was neccessary to add a way of identify the terminal by
initial title and associate this with their address to make it possible to apply rules to its
window.

- For konsole it wasn't neccesary to do much. The way of identifying its window was to add
a special matching because konsole changes a little bit the title of the window so a partial
regex matching was implemented.

- Refactor was made to rename the word class of the app to avoid confusion with window class.
Now the word used is identifier which is a more generic description that encapsules class,
title and address.